+++ tests/test_util.py
@@ -14,3 +14,12 @@ class TestUtil(unittest.TestCase):
         self.assertFalse(util.cmp_coords([1, 3, 2], [1, 2, 3]))
         self.assertFalse(util.cmp_coords([1, 2, NaN], [1, NaN, 3]))
         self.assertFalse(util.cmp_coords([1, 2, 3, 4], [1, 2, 3]))
+
+
+    def test_formatting_string(self):
+        """Test string formatting case."""
+
+        self.assertEqual(
+            util.fmt_float(9.345678939171937490173947397373e3, 15, 'decimal'),
+            '9345.678939171938'
+        )